Work on the Item Activation
Relates to work on the Item Activation code for the Player and Green Snowball. Currently it can be spammed and is not set up for auto use. There is still work to be done to fully realize the various pieces, but this is a good chunk of the work so far in a semi usable state.