Mystery Box + Green Snowball integration
Player can currently pick up a random item (well, Green Snowball currently with some placeholder text as well for the Blue Snowball) and use the item accordingly. There is still one more major fix to be done, but the overall implementation is done at the moment.

We need an AI or target dummy to test the effects out properly and ensure items work where needed. This is currently in development as of writing this, so no worries for the time being.