Implementation of the Blue Snowball's basics
The implementation of the Blue Snowball minus hit detection and position tracking (who is in first).

There was also an attempt to rework the Mystery Box system since not all items should be obtainable by first place (the Blue Snowball for example), but the script would not appear anywhere despite being attached to the game files. I decided to hold off on implementing until I had committed this work and done merge testing.