Addition of the Competition Controller
This new script is to handle most logic regarding positioning of a character in a race. Specifically if they pass someone, if they are passed, what placement they are in and so on.

It is a main component that works alongside various other pieces of content in the game when it comes to a character's current placement in the ongoing race.